state refactor
