+++ sequencer/closingsignalsmanager_test.go
@@ -13,6 +13,7 @@ import (
 	"github.com/0xPolygonHermez/zkevm-node/merkletree"
 	"github.com/0xPolygonHermez/zkevm-node/merkletree/hashdb"
 	"github.com/0xPolygonHermez/zkevm-node/state"
+	"github.com/0xPolygonHermez/zkevm-node/state/pgstatestorage"
 	"github.com/0xPolygonHermez/zkevm-node/state/runtime/executor"
 	"github.com/0xPolygonHermez/zkevm-node/test/dbutils"
 	"github.com/0xPolygonHermez/zkevm-node/test/testutils"
@@ -72,7 +73,7 @@ func setupTest(t *testing.T) {
 	eventLog := event.NewEventLog(event.Config{}, eventStorage)
 
 	localStateTree := merkletree.NewStateTree(localMtDBServiceClient)
-	localState = state.NewState(stateCfg, state.NewPostgresStorage(state.Config{}, localStateDb), localExecutorClient, localStateTree, eventLog)
+	localState = state.NewState(stateCfg, pgstatestorage.NewPostgresStorage(state.Config{}, localStateDb), localExecutorClient, localStateTree, eventLog)
 
 	batchConstraints := state.BatchConstraintsCfg{
 		MaxTxsPerBatch:       300,
@@ -110,7 +111,7 @@ func prepareForcedBatches(t *testing.T) {
 
 	for x := 0; x < numberOfForcesBatches; x++ {
 		forcedBatchNum := int64(x)
-		_, err := localState.PostgresStorage.Exec(localCtx, sql, forcedBatchNum, testGER.String(), time.Now(), testRawData, testAddr.String(), 0)
+		_, err := localState.Exec(localCtx, sql, forcedBatchNum, testGER.String(), time.Now(), testRawData, testAddr.String(), 0)
 		assert.NoError(t, err)
 	}
 }
