+++ state/batch.go
@@ -114,7 +114,7 @@ func (s *State) OpenBatch(ctx context.Context, processingContext ProcessingConte
 		return ErrDBTxNil
 	}
 	// Check if the batch that is being opened has batch num + 1 compared to the latest batch
-	lastBatchNum, err := s.PostgresStorage.GetLastBatchNumber(ctx, dbTx)
+	lastBatchNum, err := s.GetLastBatchNumber(ctx, dbTx)
 	if err != nil {
 		return err
 	}
@@ -122,7 +122,7 @@ func (s *State) OpenBatch(ctx context.Context, processingContext ProcessingConte
 		return fmt.Errorf("%w number %d, should be %d", ErrUnexpectedBatch, processingContext.BatchNumber, lastBatchNum+1)
 	}
 	// Check if last batch is closed
-	isLastBatchClosed, err := s.PostgresStorage.IsBatchClosed(ctx, lastBatchNum, dbTx)
+	isLastBatchClosed, err := s.IsBatchClosed(ctx, lastBatchNum, dbTx)
 	if err != nil {
 		return err
 	}
@@ -137,7 +137,7 @@ func (s *State) OpenBatch(ctx context.Context, processingContext ProcessingConte
 	if prevTimestamp.Unix() > processingContext.Timestamp.Unix() {
 		return ErrTimestampGE
 	}
-	return s.PostgresStorage.openBatch(ctx, processingContext, dbTx)
+	return s.OpenBatchInStorage(ctx, processingContext, dbTx)
 }
 
 // ProcessSequencerBatch is used by the sequencers to process transactions into an open batch
@@ -210,7 +210,7 @@ func (s *State) ExecuteBatch(ctx context.Context, batch Batch, updateMerkleTree 
 	}
 
 	// Get previous batch to get state root and local exit root
-	previousBatch, err := s.PostgresStorage.GetBatchByNumber(ctx, batch.BatchNumber-1, dbTx)
+	previousBatch, err := s.GetBatchByNumber(ctx, batch.BatchNumber-1, dbTx)
 	if err != nil {
 		return nil, err
 	}
@@ -278,7 +278,7 @@ func (s *State) processBatch(ctx context.Context, batchNumber uint64, batchL2Dat
 		return nil, ErrExecutorNil
 	}
 
-	lastBatches, err := s.PostgresStorage.GetLastNBatches(ctx, two, dbTx)
+	lastBatches, err := s.GetLastNBatches(ctx, two, dbTx)
 	if err != nil {
 		return nil, err
 	}
@@ -292,7 +292,7 @@ func (s *State) processBatch(ctx context.Context, batchNumber uint64, batchL2Dat
 		previousBatch = lastBatches[1]
 	}
 
-	isBatchClosed, err := s.PostgresStorage.IsBatchClosed(ctx, batchNumber, dbTx)
+	isBatchClosed, err := s.IsBatchClosed(ctx, batchNumber, dbTx)
 	if err != nil {
 		return nil, err
 	}
@@ -364,7 +364,7 @@ func (s *State) sendBatchRequestToExecutor(ctx context.Context, processBatchRequ
 
 func (s *State) isBatchClosable(ctx context.Context, receipt ProcessingReceipt, dbTx pgx.Tx) error {
 	// Check if the batch that is being closed is the last batch
-	lastBatchNum, err := s.PostgresStorage.GetLastBatchNumber(ctx, dbTx)
+	lastBatchNum, err := s.GetLastBatchNumber(ctx, dbTx)
 	if err != nil {
 		return err
 	}
@@ -372,7 +372,7 @@ func (s *State) isBatchClosable(ctx context.Context, receipt ProcessingReceipt, 
 		return fmt.Errorf("%w number %d, should be %d", ErrUnexpectedBatch, receipt.BatchNumber, lastBatchNum)
 	}
 	// Check if last batch is closed
-	isLastBatchClosed, err := s.PostgresStorage.IsBatchClosed(ctx, lastBatchNum, dbTx)
+	isLastBatchClosed, err := s.IsBatchClosed(ctx, lastBatchNum, dbTx)
 	if err != nil {
 		return err
 	}
@@ -394,7 +394,7 @@ func (s *State) CloseBatch(ctx context.Context, receipt ProcessingReceipt, dbTx 
 		return err
 	}
 
-	return s.PostgresStorage.closeBatch(ctx, receipt, dbTx)
+	return s.CloseBatchInStorage(ctx, receipt, dbTx)
 }
 
 // ProcessAndStoreClosedBatch is used by the Synchronizer to add a closed batch into the data base. Values returned are the new stateRoot,
@@ -467,7 +467,7 @@ func (s *State) ProcessAndStoreClosedBatch(ctx context.Context, processingCtx Pr
 	}
 
 	// Close batch
-	return common.BytesToHash(processed.NewStateRoot), processed.FlushId, processed.ProverId, s.closeBatch(ctx, ProcessingReceipt{
+	return common.BytesToHash(processed.NewStateRoot), processed.FlushId, processed.ProverId, s.CloseBatchInStorage(ctx, ProcessingReceipt{
 		BatchNumber:   processingCtx.BatchNumber,
 		StateRoot:     processedBatch.NewStateRoot,
 		LocalExitRoot: processedBatch.NewLocalExitRoot,
@@ -478,7 +478,7 @@ func (s *State) ProcessAndStoreClosedBatch(ctx context.Context, processingCtx Pr
 
 // GetLastBatch gets latest batch (closed or not) on the data base
 func (s *State) GetLastBatch(ctx context.Context, dbTx pgx.Tx) (*Batch, error) {
-	batches, err := s.PostgresStorage.GetLastNBatches(ctx, 1, dbTx)
+	batches, err := s.GetLastNBatches(ctx, 1, dbTx)
 	if err != nil {
 		return nil, err
 	}
