+++ state/helper_test.go
@@ -14,11 +14,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-const (
-	forkID5 = 5
-	forkID4 = 4
-)
-
 func init() {
 	log.Init(log.Config{
 		Level:   "debug",
@@ -28,19 +23,19 @@ func init() {
 
 func TestDecodeRandomBatchL2Data(t *testing.T) {
 	randomData := []byte("Random data")
-	txs, _, _, err := state.DecodeTxs(randomData, forkID5)
+	txs, _, _, err := state.DecodeTxs(randomData, state.DRAGONFRUIT_FORKID)
 	require.Error(t, err)
 	assert.Equal(t, []types.Transaction{}, txs)
 	t.Log("Txs decoded 1: ", txs)
 
 	randomData = []byte("Esto es autentica basura")
-	txs, _, _, err = state.DecodeTxs(randomData, forkID5)
+	txs, _, _, err = state.DecodeTxs(randomData, state.DRAGONFRUIT_FORKID)
 	require.Error(t, err)
 	assert.Equal(t, []types.Transaction{}, txs)
 	t.Log("Txs decoded 2: ", txs)
 
 	randomData = []byte("beef")
-	txs, _, _, err = state.DecodeTxs(randomData, forkID5)
+	txs, _, _, err = state.DecodeTxs(randomData, state.DRAGONFRUIT_FORKID)
 	require.Error(t, err)
 	assert.Equal(t, []types.Transaction{}, txs)
 	t.Log("Txs decoded 3: ", txs)
@@ -49,7 +44,7 @@ func TestDecodeRandomBatchL2Data(t *testing.T) {
 func TestDecodePre155BatchL2DataPreForkID5(t *testing.T) {
 	pre155, err := hex.DecodeString("e480843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77cb7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feb246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed1b")
 	require.NoError(t, err)
-	txs, _, _, err := state.DecodeTxs(pre155, forkID4)
+	txs, _, _, err := state.DecodeTxs(pre155, state.BLUEBERRY_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs, len(txs))
 	assert.Equal(t, 1, len(txs))
@@ -65,7 +60,7 @@ func TestDecodePre155BatchL2DataPreForkID5(t *testing.T) {
 
 	pre155, err = hex.DecodeString("e580843b9aca00830186a0941275fbb540c8efc58b812ba83b0d0b8b9917ae988084159278193d7bcd98c00060650f12c381cc2d4f4cc8abf54059aecd2c7aabcfcdd191ba6827b1e72f0eb0b8d5daae64962f4aafde7853e1c102de053edbedf066e6e3c2dc1b")
 	require.NoError(t, err)
-	txs, _, _, err = state.DecodeTxs(pre155, forkID4)
+	txs, _, _, err = state.DecodeTxs(pre155, state.BLUEBERRY_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs)
 	assert.Equal(t, 1, len(txs))
@@ -80,7 +75,7 @@ func TestDecodePre155BatchL2DataPreForkID5(t *testing.T) {
 func TestDecodePre155BatchL2DataForkID5(t *testing.T) {
 	pre155, err := hex.DecodeString("e480843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77cb7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feb246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed1bff")
 	require.NoError(t, err)
-	txs, _, _, err := state.DecodeTxs(pre155, forkID5)
+	txs, _, _, err := state.DecodeTxs(pre155, state.DRAGONFRUIT_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs, len(txs))
 	assert.Equal(t, 1, len(txs))
@@ -96,7 +91,7 @@ func TestDecodePre155BatchL2DataForkID5(t *testing.T) {
 
 	pre155, err = hex.DecodeString("e580843b9aca00830186a0941275fbb540c8efc58b812ba83b0d0b8b9917ae988084159278193d7bcd98c00060650f12c381cc2d4f4cc8abf54059aecd2c7aabcfcdd191ba6827b1e72f0eb0b8d5daae64962f4aafde7853e1c102de053edbedf066e6e3c2dc1b")
 	require.NoError(t, err)
-	txs, _, _, err = state.DecodeTxs(pre155, forkID4)
+	txs, _, _, err = state.DecodeTxs(pre155, state.BLUEBERRY_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs)
 	assert.Equal(t, 1, len(txs))
@@ -127,9 +122,9 @@ func TestDecodePre155Tx(t *testing.T) {
 func TestEncodePre155BatchL2DataPreForkID5(t *testing.T) {
 	pre155, err := hex.DecodeString("e480843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77cb7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feb246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed1b")
 	require.NoError(t, err)
-	txs, _, effectivePercentages, err := state.DecodeTxs(pre155, forkID4)
+	txs, _, effectivePercentages, err := state.DecodeTxs(pre155, state.BLUEBERRY_FORKID)
 	require.NoError(t, err)
-	rawtxs, err := state.EncodeTransactions(txs, effectivePercentages, forkID4)
+	rawtxs, err := state.EncodeTransactions(txs, effectivePercentages, state.BLUEBERRY_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs, len(txs))
 	assert.Equal(t, pre155, rawtxs)
@@ -138,9 +133,9 @@ func TestEncodePre155BatchL2DataPreForkID5(t *testing.T) {
 func TestEncodePre155BatchL2DataForkID5(t *testing.T) {
 	pre155, err := hex.DecodeString("e480843b9aca00826163941275fbb540c8efc58b812ba83b0d0b8b9917ae98808464fbb77cb7d2a666860f3c6b8f5ef96f86c7ec5562e97fd04c2e10f3755ff3a0456f9feb246df95217bf9082f84f9e40adb0049c6664a5bb4c9cbe34ab1a73e77bab26ed1bff")
 	require.NoError(t, err)
-	txs, _, effectivePercentages, err := state.DecodeTxs(pre155, forkID5)
+	txs, _, effectivePercentages, err := state.DecodeTxs(pre155, state.DRAGONFRUIT_FORKID)
 	require.NoError(t, err)
-	rawtxs, err := state.EncodeTransactions(txs, effectivePercentages, forkID5)
+	rawtxs, err := state.EncodeTransactions(txs, effectivePercentages, state.DRAGONFRUIT_FORKID)
 	require.NoError(t, err)
 	t.Log("Txs decoded: ", txs, len(txs))
 	assert.Equal(t, pre155, rawtxs)
@@ -206,7 +201,7 @@ func TestMaliciousTransaction(t *testing.T) {
 		0xd2, 0x05, 0xf0, 0xa3, 0x6f, 0xdc, 0x6e, 0x4e, 0x4c, 0x5a, 0x7b, 0x88,
 		0xd4, 0x5b, 0x1b}
 
-	_, _, _, err = state.DecodeTxs(b, forkID4)
+	_, _, _, err = state.DecodeTxs(b, state.BLUEBERRY_FORKID)
 	require.Error(t, err)
 	require.Equal(t, err, state.ErrInvalidData)
 }
